adjust file content, remove the file preprocessing.py, and combine its content to processing_pdf.py
